add new package for refactoring to easily preserve old code
